AI_Detection/requirements.txt
# This file lists the dependencies required for the AI Detection project.
torch
opencv-python
requests
paho-mqtt
numpy